Allow a per-toolchain specified pkg-config binary
When cross-compilation toolchains are used, they often provide their own
pkg-config binary, which automatically points to the right location for the
.pc files, the libraries and headers.

In an Eclipse CDT context, those toolchains are typically registered through
the org.eclipse.cdt.managedbuilder.core.buildDefinitions extension point. In
this model, the toolchain are composed of several tools: the C compiler, the
C++ compiler, the linker, the assembler, etc.

So, what we propose in this patch is to make pkg-config another "tool" of
a toolchain, so that Eclipse plugins registering toolchains can specify the
path to a toolchain-specific pkg-config binary along with the path to other
traditional tools (compilers, etc.). To do so, they would define a tool whose
superClass is org.eclipse.cdt.managedbuilder.pkgconfig.tool (this tool is
defined as an abstract tool in our pkg-config plugin).

Once we have this in place, this patch tunes the configuration interface in
terms of pkg-config path selection. We now have two choices:

 * Use the default pkg-config path provided by the toolchain. If this option
   is selected but the toolchain does not provide a pkg-config tool, then the
   default of just 'pkg-config' is assumed (or pkg-config.exe on Windows).

 * A custom path, completely user-defined.

This patch also adds a ICProjectDescriptionListener, which allows us to be
notified when the toolchain selection is changed, so that we can update the
pkg-config settings accordingly.

Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>